+++ nowplaying/artistextras/fanarttv.py
@@ -103,7 +103,7 @@ class Plugin(ArtistExtrasPlugin):
         return metadata
 
     def providerinfo(self):  # pylint: disable=no-self-use
-        ''' return list of what is provided by this recognition system '''
+        ''' return list of what is provided by this plug-in '''
         return [
             'artistbannerraw', 'artistlogoraw', 'artistthumbraw',
             'fanarttv-artistfanarturls'