+++ nowplaying/recognition/acoustidmb.py
@@ -29,7 +29,7 @@ import nowplaying.version
 
 
 class Plugin(RecognitionPlugin):
-    ''' handler for NowPlaying '''
+    ''' handler for acoustidmb '''
 
     def __init__(self, config=None, qsettings=None):
         super().__init__(config=config, qsettings=qsettings)